requirements.txt
absl-py==0.8.1
altair==3.2.0
argh==0.26.2
astor==0.8.0
attrs==19.3.0
base58==1.0.3
blinker==1.4
boto3==1.10.22
botocore==1.13.22
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.1
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
future==0.18.2
gast==0.2.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
importlib-metadata==0.23
Jinja2==2.10.3
jmespath==0.9.4
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==7.2.0
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==1.0.0
pathtools==0.1.2
Pillow==6.2.1
protobuf==3.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pydot==1.3.0
pyparsing==2.4.6
pyrsistent==0.15.5
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.2.1
six==1.14.0
streamlit==0.52.2
tensorboard==2.0.1
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.0
termcolor==1.1.0
toml==0.10.0
toolz==0.10.0
tornado==6.0.3
tzlocal==2.0.0
urllib3==1.25.7
validators==0.14.0
watchdog==0.9.0
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
